清理 Rust test-target 告警面并归档 OpenSpec 变更
