+++ openspec/changes/archive/2026-04-23-clean-rust-test-target-warning-surface/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-04-23
